Move chain-level members from Asset to Chain; rename EvmCoin to Ether
