+++ tests/test_self_transfer_guard.py
@@ -11,10 +11,11 @@ reward-side limits, so it still passes here.
 from typing import Optional
 
 from allways.assets.base import Asset, TransactionInfo
+from allways.assets.chain import Chain
 from allways.chains import CHAIN_TAO, ChainDefinition
 
 
-class FakeProvider(Asset):
+class FakeProvider(Asset, Chain):
     """Returns a single canned tx; only the base post-checks are exercised."""
 
     def __init__(self, tx: TransactionInfo):
